# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/checkin": {
            "post": {
                "description": "用户手动打卡，记录当天学习情况",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "打卡"
                ],
                "summary": "今日打卡",
                "responses": {
                    "200": {
                        "description": "打卡成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "打卡失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/checkin-stats": {
            "get": {
                "description": "获取用户的打卡天数、连续天数等统计信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "打卡"
                ],
                "summary": "获取打卡统计",
                "responses": {
                    "200": {
                        "description": "打卡统计信息",
                        "schema": {
                            "$ref": "#/definitions/models.CheckInStat"
                        }
                    },
                    "500": {
                        "description": "获取统计失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/questionbanks": {
            "get": {
                "description": "分页获取题库列表，可按类型和排序方式筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "获取题库列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "题库类型(official/shared/personal)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方式(star_count/fork_count/created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "题库列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的题库，可选导入题目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "创建题库",
                "parameters": [
                    {
                        "description": "题库信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "创建失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/questionbanks-progress": {
            "get": {
                "description": "获取当前用户所有题库的学习进度信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "获取所有题库的学习进度",
                "responses": {
                    "200": {
                        "description": "进度信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/questionbanks/{id}": {
            "put": {
                "description": "更新指定ID的题库信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "更新题库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "题库信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "更新失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定ID的题库",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "删除题库",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "题库不存在或无权限",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "题库被学习计划使用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/questionbanks/{id}/progress": {
            "get": {
                "description": "获取指定题库的学习进度信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "题库"
                ],
                "summary": "获取特定题库的学习进度",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "题库ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "进度信息",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/study-heatmap": {
            "get": {
                "description": "获取用户一年内的学习热力图数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "打卡"
                ],
                "summary": "获取学习热力图",
                "responses": {
                    "200": {
                        "description": "热力图数据",
                        "schema": {
                            "$ref": "#/definitions/services.HeatmapResponse"
                        }
                    },
                    "500": {
                        "description": "获取热力图失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CheckInStat": {
            "type": "object",
            "properties": {
                "best_streak": {
                    "description": "历史最长连续天数",
                    "type": "integer"
                },
                "checked_in_today": {
                    "description": "今日是否已打卡",
                    "type": "boolean"
                },
                "consecutive_days": {
                    "description": "当前连续打卡天数",
                    "type": "integer"
                },
                "total_check_in_days": {
                    "description": "总打卡天数",
                    "type": "integer"
                }
            }
        },
        "services.HeatmapData": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "当天学习题目数量",
                    "type": "integer"
                },
                "date": {
                    "description": "YYYY-MM-DD 格式",
                    "type": "string"
                },
                "level": {
                    "description": "活跃度级别 0-4",
                    "type": "integer"
                }
            }
        },
        "services.HeatmapResponse": {
            "type": "object",
            "properties": {
                "current_streak": {
                    "description": "当前连续天数",
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.HeatmapData"
                    }
                },
                "max_streak": {
                    "description": "最长连续天数",
                    "type": "integer"
                },
                "this_year": {
                    "description": "今年学习天数",
                    "type": "integer"
                },
                "total_commits": {
                    "description": "总学习天数",
                    "type": "integer"
                }
            }
        }
    }
}